course stuff

+++ Responsible Student Simulator/src/Course.java
@@ -62,6 +62,6 @@ public class Course {
 	}
 	
 	public double getProgress() {
-		return productionProgress / productionFinish;
+		return (double)productionProgress / productionFinish;
 	}
 }
